# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Agent Deployment System — Control Plane API (Phase 1-2)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/{tenantId}/workspaces/{workspaceId}/environments:
    post:
      operationId: createEnvironment
      tags: [environments]
      summary: Create an environment within a workspace
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironmentBody"
      responses:
        "201":
          description: Environment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: listEnvironments
      tags: [environments]
      summary: List environments for a workspace
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - $ref: "#/components/parameters/workspaceId"
        - name: type
          in: query
          schema:
            type: string
            enum: [development, staging, production]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of environments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentList"
  /tenants/{tenantId}/workspaces/{workspaceId}/environments/{environmentId}:
    get:
      operationId: getEnvironment
      tags: [environments]
      summary: Get an environment
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/environmentId"
      responses:
        "200":
          description: Environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateEnvironment
      tags: [environments]
      summary: Update an environment
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/environmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnvironmentBody"
      responses:
        "200":
          description: Updated environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── Packages ──────────────────────────────────────────────────────────────
...
  /tenants/{tenantId}/approvals:
    post:
      operationId: createApprovalRequest
      tags: [approvals]
      summary: Create an approval request
      parameters:
        - $ref: "#/components/parameters/tenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRequestBody"
      responses:
        "201":
          description: Approval request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
    get:
      operationId: listApprovalRequests
      tags: [approvals]
      summary: List approval requests for a tenant
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, expired]
        - name: requesterId
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of approval requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequestList"
  /tenants/{tenantId}/approvals/{approvalId}:
    get:
      operationId: getApprovalRequest
      tags: [approvals]
      summary: Get an approval request
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - $ref: "#/components/parameters/approvalId"
      responses:
        "200":
          description: Approval request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /tenants/{tenantId}/approvals/{approvalId}/decision:
    post:
      operationId: submitApprovalDecision
      tags: [approvals]
      summary: Submit an approval or rejection decision
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - $ref: "#/components/parameters/approvalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionBody"
      responses:
        "200":
          description: Updated approval request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Decision already made
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── Policy (default-deny skeleton) ───────────────────────────────────────
...
components:
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
    environmentId:
      name: environmentId
      in: path
      required: true
      schema:
        type: string
...
    approvalId:
      name: approvalId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
    Metadata:
      type: object
      additionalProperties: true

    # ── Tenant ────────────────────────────────────────────────────────────────
...
    Environment:
      type: object
      properties:
        id:
          type: string
          description: "Prefixed ID: env_{nanoid}"
        workspaceId:
          type: string
        tenantId:
          type: string
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [development, staging, production]
        status:
          type: string
          enum: [active, inactive]
        metadata:
          $ref: "#/components/schemas/Metadata"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, workspaceId, tenantId, name, slug, type, status, createdAt, updatedAt]
    EnvironmentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Environment"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
    CreateEnvironmentBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [development, staging, production]
        metadata:
          $ref: "#/components/schemas/Metadata"
      required: [name, slug, type]
    UpdateEnvironmentBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, inactive]
        metadata:
          $ref: "#/components/schemas/Metadata"

    # ── Package ───────────────────────────────────────────────────────────────
...
    ApprovalRequest:
      type: object
      properties:
        id:
          type: string
          description: "Prefixed ID: apr_{nanoid}"
        tenantId:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
        action:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        requesterId:
          type: string
        reviewerId:
          type: string
          nullable: true
        requestPayload:
          type: object
          additionalProperties: true
          nullable: true
        decisionPayload:
          type: object
          additionalProperties: true
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        decidedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tenantId, resourceType, resourceId, action, status, requesterId, createdAt, updatedAt]
    ApprovalRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequest"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
    CreateApprovalRequestBody:
      type: object
      properties:
        resourceType:
          type: string
        resourceId:
          type: string
        action:
          type: string
        requesterId:
          type: string
        requestPayload:
          type: object
          additionalProperties: true
        expiresAt:
          type: string
          format: date-time
      required: [resourceType, resourceId, action, requesterId]
    ApprovalDecisionBody:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        reviewerId:
          type: string
        decisionPayload:
          type: object
          additionalProperties: true
      required: [decision, reviewerId]

    # ── Policy ────────────────────────────────────────────────────────────────
...